server-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
coinbase-advanced-py==1.8.2
cryptography>=42.0.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
# mcp==1.0.0  # Will install later if needed